Enable term toggling
It isn't the best toggling solution but it's easy and works most of the
time. The problem is that sometimes toggling the terminal open will
trigger execution of the command, which defeats the purpose of toggling
open the terminal to see the `make` output (because it just starts
`make`ing again!). But, it seems to work fine for extant processes and
even most of the time for completed processes (weird?). I _think_ what
is happening is that you cannot toggle it open twice, because internally
toggleterm has set up some autocommands to delete the terminal when it
closes (which is seemingly *only* after the process has completed?). So
this is something I will likely submit a PR for, along with the ability
to configure a per terminal "start_in_insert" option.